update alireza files to reflect renaming of data files
